Disabled reading log on return to page 1
The future was introduced earlier to avoid "hallucinations" after not being
on page 1 for a long time (time ticks but the logs are not updated when you
are on other pages).

But if you return to page 1 and it does not have new log information it's not
too bad - it will look like you are having a major NFS freeze but that will
update within 30 seconds. It's much worse having the 5 second freeze every time
you return to page 1.